requirements.txt
SQLAlchemy
requests
redis
alembic
python-dotenv
psycopg2
uvicorn
fastapi